src/requirements.txt
absl-py==1.4.0
asttokens==2.1.0
by==0.0.4
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
Cython==0.29.35
debugpy==1.6.3
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fasttext==0.9.2
folium==0.13.0
fonttools==4.33.3
fsspec==2022.8.2
future==0.18.3
gast==0.4.0
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
hgtk==0.2.0
huggingface-hub==0.13.4
iniconfig==2.0.0
ipython==8.6.0
jax==0.4.10
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
JPype1==1.4.1
Js2Py==0.74
jupyter_client==7.4.6
jupyter_core==5.0.0
keras==2.12.0
kiwisolver==1.4.2
konlpy==0.6.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.22.4
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.4
pluggy==1.0.0
prompt-toolkit==3.0.32
protobuf==4.23.2
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pybind11==2.10.4
pycparser==2.21
Pygments==2.13.0
pyjsparser==2.7.1
pypandoc==1.11
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
PySocks==1.7.1
pytest==7.3.0
python-dateutil==2.8.2
pytz==2022.4
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.8.1
selenium==4.9.0
seungminpark-calc==0.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soynlp==0.0.493
SpeechRecognition==3.10.0
stack-data==0.6.1
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboardX==2.6
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
tornado==6.2
tqdm==4.65.0
trainer==0.0.24
traitlets==5.5.0
transformers==4.27.4